Migrate validation tests to test_structure package
